Refactor: move methods to god.cpp
