+++ variables.hpp
@@ -796,7 +796,6 @@ int map_placeupstairs(int = 0, int = 0);
 int map_trap(int = 0, int = 0, int = 0, int = 0);
 int map_web(int = 0, int = 0, int = 0);
 int mapitemfind(int = 0, int = 0, int = 0);
-int modpiety(int = 0);
 int net_dl(const std::string&, const std::string&);
 int net_dllist(const std::string&, int = 0);
 int net_read(int = 0);
@@ -1068,13 +1067,7 @@ void label_1877();
 void casino_wrapper();
 bool casino_start();
 bool casino_blackjack();
-void set_npc_religion();
-void apply_god_blessing(int);
-void get_god_description();
 void label_1886();
-void label_1888();
-void switch_religion();
-turn_result_t do_pray();
 void label_1892();
 void start_stealing();
 void label_1901();
